more refactor and improvements for activation and login
