+++ packages/client/src/pages/Activate/Screens/ActivateSreensStyles.tsx
