+++ packages/client/src/pages/Login/Screens/PasswordRecoverScreen.tsx
@@ -2,18 +2,24 @@ import { config, useSpring } from "@react-spring/web";
 import { Button, Input } from "components";
 import React, { useState } from "react";
 import { IoReloadCircle } from "react-icons/io5";
-import {
-  StyledInputRow,
-  StyledButtonWrap,
-  StyledErrorText,
-  StyledTbMailFilled,
-} from "../LoginModalStyles";
+// import {
+//   StyledInputRow,
+//   StyledButtonWrap,
+//   StyledErrorText,
+//   StyledTbMailFilled,
+// } from "../LoginModal/LoginModalStyles";
 import {
   StyledAnimatedIconWrap,
   StyledDescription,
   StyledEmailSent,
   StyledIoReloadCircle,
-} from "./ScreensStyles";
+  StyledTbMailFilled,
+} from "./LoginScreensStyles";
+import {
+  StyledButtonWrap,
+  StyledErrorText,
+  StyledInputRow,
+} from "pages/PasswordReset/PasswordResetPageStyles";
 
 interface PasswordRecoverScreen {
   emailLocal: string;
